pyproject.toml
[project]
name = "saturnus-magister"
version = "1.0.0"
requires-python = ">=3.14"
description = "Autonomous email orchestration system syncing Gmail, Grok classification, Nyx_Venatrix job matching, and TickTick routing for intelligent task management"

dependencies = [
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "asyncpg>=0.30",
    "httpx>=0.28",
    "aiofiles>=24.1",
    "google-api-python-client>=2.150",
    "google-auth-oauthlib>=1.2",
    "openai>=1.55",
    "rapidfuzz>=3.10",
    "rich>=13.9",
    "typer>=0.13",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
saturnus = "src.main:cli"
saturnus-review = "src.cli.review:main"
saturnus-setup = "src.cli.setup:main"

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.mypy]
python_version = "3.14"
strict = true
